core: snap train speed to speed constraints
this makes it so that when the rolling stock reaches a speed close to a
limit (either the current speed limit or one of the deceleration curve),
the integration step is fixed to "snap" the speed to the limit.

The threshold of closeness is 1e-4 (chosen arbitrarily)

I think this is needed because of floating point errors, but also
because deceleration curves are interpolated as segmented lines (instead
of a quadratic curve), although i haven't tested if changing how we
interpolate them fixes the problem.

Additionally, the time step may be decreased to improve correctness.